scraper_requirement.txt
selenium
fake-useragent
requests
beautifulsoup4
pandas
pymongo
python-dotenv